Solved a bug in the Supervised Learning code, to allow to train with float16 precision
Next time, we will train AlphaZero with less batchsize, float16 and smaller network, to see if the train time is near the 2 weeks

+++ AlphaZero/src/train/AlphaZeroV2.cpp
@@ -59,7 +59,7 @@ AlphaZeroV2::AlphaZeroV2(
     }
     else if (precision_type == "float16")
     {
-        precision = torch::kFloat16;
+        precision = torch::kBFloat16;
     }
     else
     {
